+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## 0.33.0 - 2026-04-27
+### Added
+- prompt wallet to add missing selected chain
+
+### Fixed
+- fix token swap in step 1 to keep metadata and resolved addresses in sync
 
 ## 0.32.0 - 2026-04-27
 ### Added